+++ src/lib.rs
@@ -8,6 +8,7 @@ mod client;
 mod http_client;
 mod record;
 mod replication;
+mod token;
 
 pub use bucket::Bucket;
 pub use client::ReductClient;
@@ -33,3 +34,4 @@ pub use reduct_base::msg::token_api::{Permissions, Token};
 pub use serde_json::json as condition;
 pub use serde_json::json as ext; // for readability
 pub use serde_json::Value as JsonValue;
+pub use token::{TokenCreateOptions, TokenInfo, TokenInfoList};